+++ _site/_includes/components/in-page-nav/whole-page-nav.js
@@ -1,1 +1,0 @@
-$('#demo-whole-page-nav-tabs').inPageNav();