+++ _site/_includes/components/lightbox/small.js
@@ -1,1 +1,0 @@
-$('#lightbox-small-demo-link').lightbox({size:'small'});