+++ grunt/aliases.js
@@ -1,5 +1,5 @@
 module.exports = {
-    "dev-build": ['clean', 'compass', 'requirejs', 'fonts', 'svgs', 'uglify', 'jekyll:build', 'prepare-site'],
+    "dev-build": ['clean', 'exec:bower-install', 'compass', 'requirejs', 'svgs', 'uglify', 'jekyll:build', 'prepare-site'],
 
     //  standard build tasks that lints your JS
     "build": ['jshint', 'version_sync', 'dev-build'],
@@ -7,11 +7,10 @@ module.exports = {
     "build-js": ['jshint', 'requirejs', 'uglify'],
 
     //  misc tasks
-    "fonts": ['clean:fonts-min', 'clean:fonts', 'svgmin:fonts', 'webfont', 'cssmin:skycons'],
     "svgs": ['clean:svgs-min', 'clean:svgs', 'svgmin:icons', 'grunticon'],
 
     // for toolkit website with runnable tests
-    "prepare-site": ['copy:dist', 'copy:test'],
+    "prepare-site": ['copy:bower_fonts', 'copy:dist', 'copy:test'],
 
     //  testing tasks
     "test-with-coverage": ['blanket_mocha'],