+++ test/specs/global-spec.js
@@ -2,7 +2,7 @@ describe('global (non-AMD) toolkit', function () {
 
     it('should not throw an error', function () {
         expect(toolkit).to.not.equal(undefined);
-        expect(toolkit.detect).to.not.equal(undefined);
+        expect(skyComponents.detect).to.not.equal(undefined);
     });
 
 });