+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Change Log
 
+## 2.3.4
+
+  * Detect
+   * [updated] Now using the [Detect component](https://github.com/skyglobal/detect)
+  * Event
+   * [updated] Now using the [Event component](https://github.com/skyglobal/event)
+
+## 2.3.3
+ * Polyfill
+   * [updated] Now using the [Polyfill component](https://github.com/skyglobal/polyfill)
+ * Skycons
+   * [updated] Now using the [Skycons component](https://github.com/skyglobal/skycons)
+ * Colours
+   * [updated] Now using the [Colours component](https://github.com/skyglobal/colours)
+ * Typography
+   * [updated] Now using the [Typography component](https://github.com/skyglobal/typography)
+ * Buttons
+   * [updated] Now using the [Buttons component](https://github.com/skyglobal/buttons)
+ * Grid
+   * [updated] Now using the [Grid component](https://github.com/skyglobal/grid)
+ * Spinners
+   * [updated] Now using the [Spinners component](https://github.com/skyglobal/spinners)
+   
+## 2.3.2
+ * Carousel
+   * [added] Capability to specify which slide to start the carousel with
+
+## 2.3.1
+ * Lightbox
+   * [added] Support for FORM submission
+
 ## 2.3.0 - breaking change
  * Skycons
    * [updated] Rendering engine updated (to 'node') meaning slight changes might be spotted.
@@ -41,7 +72,7 @@
 
   * Libraries
     * [added] RequireJS version 2.1.14 (demo page upgraded to use 2.1.14 from 2.1.9)
-  * Tooltip 
+  * Tooltip
     * [fixed] Now shows on top of hover area if any of the tooltip area is off the screen
   * Toolkit
     * [added] Toolkit can now be loaded before the body tag (e.g., in the head tag) without any JS errors